Issue #30: "Memory leak?" resolved. It wasn't strictly a memory leak, but primarily a logic error. When Pedestrians are moving around the map, they "register" with the tilemap block that they are in, essentially saying, "I'm here," and as other Pedestrians pass by they use this information to efficiently figure out who is nearby, and therefore, which Pedestrians for which they need to run collision checks. The problem was twofold: 1. Pedestrians were registering with the same tilemap block multiple times, so collision checks ran against the same Pedestrians over and over in a single pass 2. Pedestrians were never unregistering from tilemap blocks that they no longer occupied, so they appeared to be there forever
This quickly lead to a situation where every Pedestrian was collision checking against every other Pedestrian, and the system was filled MOSTLY with duplicate collisioin checks between the same two Pedestrians. This meant there was an ever increasing, and quickly grown, amount of calculation overhead. It also explains why we saw both a very fast increase in memory usage, and such a performance loss:
1. The number of collision checks required per logic update was increasing at an exponential rate over time (performance hit)
2. The number of registered Pedestrians on a given tile was growing at a linear rate, and never decreasing (increased memory usage)

Given this situation, it's surprising that the simulation actually ran as well as it did. With this bug in place, we were probably seeing tens of thousands of collision checks required PER LOGIC UPDATE, for just 100 Pedestrians, within 30 seconds or so.